# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Stock Money Flow Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /institutional:
    get:
      operationId: listInstitutional
      tags: [institutional]
      summary: List recent 13F institutional position changes
      parameters:
        - name: ticker
          in: query
          schema:
            type: string
        - name: action
          in: query
          schema:
            type: string
            enum: [buy, sell, increase, decrease]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Institutional transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstitutionalPage"
  /institutional/top-movers:
    get:
      operationId: getInstitutionalTopMovers
      tags: [institutional]
      summary: Institutions with the largest net position changes this quarter
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Top institutional movers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InstitutionalMover"
  /institutional/most-bought:
    get:
      operationId: getInstitutionalMostBought
      tags: [institutional]
      summary: Most bought tickers by institutions this quarter
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Most bought tickers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TickerCount"

  # ─── Insiders ─────────────────────────────────────────────────────────────
  /insiders:
    get:
      operationId: listInsiders
      tags: [insiders]
      summary: List recent SEC Form 4 insider transactions
      parameters:
        - name: ticker
          in: query
          schema:
            type: string
        - name: transactionType
          in: query
          schema:
            type: string
            enum: [buy, sell, option_exercise, gift]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Insider transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsiderPage"
...
  /insiders/notable:
    get:
      operationId: getNotableInsiders
      tags: [insiders]
      summary: Largest dollar-value insider transactions recently
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Notable insider trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InsiderTransaction"

  # ─── Politicians ──────────────────────────────────────────────────────────
  /politicians:
    get:
      operationId: listPoliticians
      tags: [politicians]
      summary: List recent congressional stock trades
      parameters:
        - name: ticker
          in: query
          schema:
            type: string
        - name: party
          in: query
          schema:
            type: string
            enum: [democrat, republican, independent]
        - name: transactionType
          in: query
          schema:
            type: string
            enum: [buy, sell]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Political trades
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoliticianPage"
  /politicians/leaders:
    get:
      operationId: getPoliticianLeaders
      tags: [politicians]
      summary: Most active trading politicians by volume
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Most active politicians
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PoliticianLeader"
  /politicians/most-traded:
    get:
      operationId: getPoliticiansMostTraded
      tags: [politicians]
      summary: Most traded tickers by politicians
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Most traded tickers by politicians
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TickerCount"

  # ─── Lockups ──────────────────────────────────────────────────────────────
  /lockups:
    get:
      operationId: listLockups
      tags: [lockups]
      summary: Upcoming and recent IPO lockup expirations
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, expired, all]
            default: upcoming
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Lockup events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LockupPage"
  /lockups/expiring-soon:
    get:
      operationId: getLockupsExpiringSoon
      tags: [lockups]
      summary: Lockups expiring in the next 30 days
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Expiring lockups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LockupEvent"
components:
  schemas:
...
    InstitutionalTransaction:
      type: object
      required: [id, institutionName, ticker, companyName, action, sharesChange, valueChange, reportDate, filingDate, quarter]
      properties:
        id:
          type: integer
        institutionName:
          type: string
        ticker:
          type: string
        companyName:
          type: string
        action:
          type: string
          enum: [buy, sell, increase, decrease]
        sharesChange:
          type: number
        valueChange:
          type: number
        totalValue:
          type: ["number", "null"]
        percentChange:
          type: ["number", "null"]
        reportDate:
          type: string
        filingDate:
          type: string
        quarter:
          type: string
    InstitutionalPage:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/InstitutionalTransaction"
        total:
          type: integer
    InstitutionalMover:
      type: object
      required: [institutionName, netValueChange, buysCount, sellsCount, topBuy, topSell]
      properties:
        institutionName:
          type: string
        netValueChange:
          type: number
        buysCount:
          type: integer
        sellsCount:
          type: integer
        topBuy:
          type: ["string", "null"]
        topSell:
          type: ["string", "null"]
    TickerCount:
      type: object
      required: [ticker, companyName, count, totalValue]
      properties:
        ticker:
          type: string
        companyName:
          type: string
        count:
          type: integer
        totalValue:
          type: number

    # ─── Insider schemas ───────────────────────────────────────────────────
    InsiderTransaction:
      type: object
      required: [id, insiderName, insiderTitle, ticker, companyName, transactionType, shares, pricePerShare, totalValue, transactionDate, filingDate]
      properties:
        id:
          type: integer
        insiderName:
          type: string
        insiderTitle:
          type: string
        ticker:
          type: string
        companyName:
          type: string
        transactionType:
          type: string
          enum: [buy, sell, option_exercise, gift]
        shares:
          type: number
        pricePerShare:
          type: number
        totalValue:
          type: number
        transactionDate:
          type: string
        filingDate:
          type: string
        isDirectOwnership:
          type: boolean
    InsiderPage:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/InsiderTransaction"
        total:
          type: integer
...
    PoliticianTrade:
      type: object
      required: [id, politician, party, chamber, ticker, companyName, transactionType, amountRange, transactionDate, disclosureDate]
      properties:
        id:
          type: integer
        politician:
          type: string
        party:
          type: string
          enum: [democrat, republican, independent]
        chamber:
          type: string
          enum: [senate, house]
        ticker:
          type: string
        companyName:
          type: string
        transactionType:
          type: string
          enum: [buy, sell]
        amountRange:
          type: string
        estimatedAmount:
          type: ["number", "null"]
        transactionDate:
          type: string
        disclosureDate:
          type: string
        daysToDisclose:
          type: integer
    PoliticianPage:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/PoliticianTrade"
        total:
          type: integer
    PoliticianLeader:
      type: object
      required: [politician, party, chamber, tradeCount, estimatedVolume, topTicker]
      properties:
        politician:
          type: string
        party:
          type: string
        chamber:
          type: string
        tradeCount:
          type: integer
        estimatedVolume:
          type: number
        topTicker:
          type: ["string", "null"]

    # ─── Lockup schemas ────────────────────────────────────────────────────
    LockupEvent:
      type: object
      required: [id, ticker, companyName, expirationDate, sharesUnlocking, estimatedValue, ipoDate, status]
      properties:
        id:
          type: integer
        ticker:
          type: string
        companyName:
          type: string
        expirationDate:
          type: string
        sharesUnlocking:
          type: number
        estimatedValue:
          type: number
        ipoDate:
          type: string
        status:
          type: string
          enum: [upcoming, expired]
        daysUntilExpiry:
          type: ["integer", "null"]
    LockupPage:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/LockupEvent"
        total:
          type: integer