# lib/api-zod/src/generated/types/politicianLeader.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Stock Money Flow Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface PoliticianLeader {
  politician: string;
  party: string;
  chamber: string;
  tradeCount: number;
  estimatedVolume: number;
  /** @nullable */
  topTicker: string | null;
}