# lib/api-zod/src/generated/types/politicianTrade.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Stock Money Flow Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { PoliticianTradeChamber } from './politicianTradeChamber';
import type { PoliticianTradeParty } from './politicianTradeParty';
import type { PoliticianTradeTransactionType } from './politicianTradeTransactionType';

export interface PoliticianTrade {
  id: number;
  politician: string;
  party: PoliticianTradeParty;
  chamber: PoliticianTradeChamber;
  ticker: string;
  companyName: string;
  transactionType: PoliticianTradeTransactionType;
  amountRange: string;
  /** @nullable */
  estimatedAmount?: number | null;
  transactionDate: string;
  disclosureDate: string;
  daysToDisclose?: number;
}